@@ -1,50 +1,39 @@
 {
-  "name": "phil-traeger",
-  "version": "0.0.0",
+  "name": "nuxt-app",
   "private": true,
+  "type": "module",
   "scripts": {
-    "dev": "vite --host",
-    "prod": "serve -s dist",
-    "build": "run-p type-check build-only",
-    "preview": "vite preview",
-    "test:unit": "vitest",
-    "build-only": "vite build",
-    "type-check": "vue-tsc --noEmit -p tsconfig.vitest.json --composite false",
-    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix --ignore-path .gitignore",
-    "format": "prettier --write src/ index.html",
-    "git": "run-p type-check format lint build"
-  },
-  "dependencies": {
-    "eslint-plugin-tailwindcss": "^3.13.0",
-    "oh-vue-icons": "^1.0.0-rc3",
-    "pinia": "^2.1.3",
-    "vue": "^3.3.4",
-    "vue-i18n": "^9.3.0-beta.24",
-    "vue-router": "^4.2.2"
+    "build": "nuxt build",
+    "dev": "nuxt dev --host",
+    "generate": "nuxt generate",
+    "preview": "nuxt preview",
+    "postinstall": "nuxt prepare",
+    "lint:js": "eslint --ext \".ts,.vue\" --ignore-path .gitignore .",
+    "lint:prettier": "prettier --check .",
+    "lint": "npm run lint:js && npm run lint:prettier",
+    "lintfix": "prettier --write --list-different . && npm run lint:js --fix"
   },
   "devDependencies": {
-    "@rushstack/eslint-patch": "^1.2.0",
-    "@tailwindcss/typography": "^0.5.9",
-    "@tsconfig/node18": "^2.0.1",
-    "@types/jsdom": "^21.1.1",
-    "@types/node": "^18.16.17",
-    "@vitejs/plugin-vue": "^4.2.3",
-    "@vue/eslint-config-prettier": "^7.1.0",
-    "@vue/eslint-config-typescript": "^11.0.3",
-    "@vue/test-utils": "^2.3.2",
-    "@vue/tsconfig": "^0.4.0",
-    "autoprefixer": "^10.4.14",
-    "daisyui": "^3.2.1",
-    "eslint": "^8.39.0",
-    "eslint-plugin-vue": "^9.11.0",
-    "jsdom": "^22.1.0",
-    "npm-run-all": "^4.1.5",
-    "postcss": "^8.4.25",
-    "prettier": "^2.8.8",
-    "tailwindcss": "^3.3.2",
-    "typescript": "~5.0.4",
-    "vite": "^4.3.9",
-    "vitest": "^0.32.0",
-    "vue-tsc": "^1.6.5"
+    "@nuxt/devtools": "latest",
+    "@nuxtjs/i18n": "^8.0.0",
+    "@tailwindcss/typography": "^0.5.13",
+    "autoprefixer": "^10.4.16",
+    "daisyui": "^4.4.23",
+    "nuxt": "^3.11.2",
+    "nuxt-icon": "^0.6.7",
+    "postcss": "^8.4.32",
+    "tailwindcss": "^3.4.0",
+    "vue": "^3.3.12",
+    "vue-router": "^4.2.5"
+  },
+  "dependencies": {
+    "@nuxt/content": "^2.11.0",
+    "@nuxt/image": "^1.3.0",
+    "@nuxtjs/eslint-config-typescript": "^12.1.0",
+    "@typescript-eslint/parser": "^6.20.0",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "prettier": "^3.2.4"
   }
 }
