@@ -1,16 +1,14 @@
-/* eslint-env node */
-require('@rushstack/eslint-patch/modern-module-resolution')
-
 module.exports = {
   root: true,
-  'extends': [
-    'plugin:vue/vue3-essential',
-    'eslint:recommended',
-    '@vue/eslint-config-typescript',
-    '@vue/eslint-config-prettier/skip-formatting',
-    'plugin:tailwindcss/recommended',
-  ],
+  env: {
+    browser: true,
+    node: true
+  },
+  parser: 'vue-eslint-parser',
   parserOptions: {
-    ecmaVersion: 'latest'
-  }
+    parser: '@typescript-eslint/parser'
+  },
+  extends: ['@nuxtjs/eslint-config-typescript', 'plugin:prettier/recommended'],
+  plugins: [],
+  rules: {}
 }
